chore(ai): the preflight announce takes its own file under the length cap
